fix: do not clean nodes with `class` (or `id` or `name`) matching "subscribe"
In BLOX CMS each article `p` is wrapped in a `div` that has a class that matches
`subscribe`. Since these comprise the article, they do not contain the article,
and are all removed in
cleaning.

Alternatively, a test could be added to check whether
nodes to be cleaned are children of the article node.

See also https://www.bloxdigital.com/solutions/content/